created separate component for the add fefault items form
