+++ client/src/pages/UserProfile/UserProfile.jsx
@@ -9,15 +9,13 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import {
   faCircleInfo,
   faMinus,
-  faUser,
-  faEnvelope,
-  faLock,
-  faKey,
-  faSave,
+
 } from '@fortawesome/free-solid-svg-icons';
 import { Container, Image } from 'react-bootstrap';
 
 import avatar from '../../assets/Anya.png';
+
+import AddDefaultItems from '../../components/AddDefaultItems/AddDefaultItems';
 import RecentCardWidget from '../../components/cardWidget/RecentCardWidget';
 import ExpensesSummary from '../ExpensesSummary/ExpensesSummary';
 import ExpenseInputFields from '../../components/addExpenseInline/ExpenseInputFields';
@@ -29,6 +27,7 @@ const UserProfile = () => {
   const currency = useSelector((state) => state.currency);
 
   const userAvatar = avatar || user?.avatar;
+  const backendLink = import.meta.env.VITE_APP_API_URL;
 
   const [profileEditMode, setProfileEditMode] = useState(false);
   const [defaultItemsEditMode, setDefaultItemsEditMode] = useState(false);
@@ -41,9 +40,7 @@ const UserProfile = () => {
 
   const [errMsgPassword, setErrMsgPassword] = useState(null);
   const [isLoading, setIsLoading] = useState(false);
-  const [defaultItems, setDefaultItems] = useState([]);
   const [income, setIncome] = useState();
-  const backendLink = import.meta.env.VITE_APP_API_URL;
   const [formData, setFormData] = useState({
     newName: '',
     newMail: '',
@@ -130,78 +127,7 @@ const UserProfile = () => {
     return true;
   };
 
-  const handleAddItem = async (newItemAdded) => {
-    const updatedBlocks = defaultItems.map((block) =>
-      block.name === newItemAdded.category
-        ? { ...block, items: [...block.items, newItemAdded] }
-        : block
-    );
-    setDefaultItems(updatedBlocks);
 
-    try {
-      const updateData = {};
-      updatedBlocks.forEach((block, index) => {
-        switch (block.name) {
-          case 'The Non-negotiables':
-            updateData.fixedItems = block.items;
-            break;
-          case 'On Repeat':
-            updateData.subscriptionItems = block.items;
-            break;
-          case 'Little Life Things':
-            updateData.otherItems = block.items;
-            break;
-          case 'Out & About':
-            updateData.transportItems = block.items;
-            break;
-          case 'Bits & Bites':
-            updateData.foodItems = block.items;
-            break;
-        }
-      });
-    } catch (error) {
-      console.error('Error fetching default items:', error);
-    }
-  };
-
-  const handleDeleteItem = async (e, blockName, indexToDelete) => {
-    const updatedBlocks = defaultItems.map((block) =>
-      block.name === blockName
-        ? {
-            ...block,
-            items: block.items.filter((_, index) => index !== indexToDelete),
-          }
-        : block
-    );
-
-    setDefaultItems(updatedBlocks);
-
-    try {
-      const updateData = {};
-
-      updatedBlocks.forEach((block, index) => {
-        switch (block.name) {
-          case 'The Non-negotiables':
-            updateData.fixedItems = block.items;
-            break;
-          case 'On Repeat':
-            updateData.subscriptionItems = block.items;
-            break;
-          case 'Little Life Things':
-            updateData.otherItems = block.items;
-            break;
-          case 'Out & About':
-            updateData.transportItems = block.items;
-            break;
-          case 'Bits & Bites':
-            updateData.foodItems = block.items;
-            break;
-        }
-      });
-    } catch (error) {
-      console.error('Error fetching card:', error);
-    }
-  };
 
   const handleSave = async () => {
     if (!validatePassword(password, password2)) {
@@ -374,64 +300,9 @@ const UserProfile = () => {
     </div>
   );
 
-  // Render default items editor
-  const renderDefaultItemsEditor = () => (
-    <div className="default-items-div group-column">
-      <div className="d-flex align-items-center justify-content-between">
-        <h5 className="mb-0">Total Income</h5>
-        <input
-          className="form-control line-input"
-          onChange={(e) => setIncome(e.target.value)}
-          value={income}
-          type="number"
-          style={{ maxWidth: '200px' }}
-        />
-      </div>
-      {defaultItems.map((group, index) => (
-        <>
-          <div className="default-items-group">
-            <h5>{group.name}</h5>
-            {group.items.map((element, index) => (
-              <div
-                className="item mb-3 month-card-body multi-column"
-                key={element.id}
-              >
-                <ul className="list-unstyled month-card-item">
-                  <li className="month-card-item-description">
-                    {element.description}
-                  </li>
-                  <li className="month-card-item-money ">
-                    {element.price} {currency}
-                    <button
-                      onClick={(e) => handleDeleteItem(e, group.name, index)}
-                      className="delete-list-item-btn"
-                    >
-                      <FontAwesomeIcon icon={faMinus} />{' '}
-                    </button>
-                  </li>
-                </ul>
-              </div>
-            ))}
-            <ExpenseInputFields
-              blockName={group.name}
-              onAdd={(newItemAdded) => handleAddItem(newItemAdded)}
-              isSmall={true}
-            />
-          </div>
-        </>
-      ))}
-      <button
-        className="save-default-item btn btn-dark w-100"
-        onClick={handleSave}
-      >
-        Save
-      </button>
-    </div>
-  );
-
   const renderDynamicContent = () => {
     if (defaultItemsEditMode) {
-      return renderDefaultItemsEditor();
+      return <AddDefaultItems />
     }
     if (profileEditMode) {
       return renderProfileEditForm();
